% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{apply_standardizer}
\alias{apply_standardizer}
\title{Apply standardization using stored center/scale}
\usage{
apply_standardizer(x, model)
}
\arguments{
\item{x}{Numeric matrix.}

\item{model}{List with center and scale vectors.}
}
\value{
Standardized matrix.
}
\description{
Apply standardization using stored center/scale
}
\keyword{internal}
