% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-survival.R
\name{build_u_grid}
\alias{build_u_grid}
\title{Build a uniform grid from 0 to tau}
\usage{
build_u_grid(tau = NULL, n = 80L)
}
\arguments{
\item{tau}{Numeric. Maximum time.}

\item{n}{Integer. Number of grid points.}
}
\value{
Numeric vector.
}
\description{
Build a uniform grid from 0 to tau
}
\keyword{internal}
