% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_outliers.R
\name{correct_outliers}
\alias{correct_outliers}
\title{Correcting individual outliers}
\usage{
correct_outliers(x, col)
}
\arguments{
\item{x}{A data.frame with the data including the .outlier column}

\item{col}{The column to correct outliers from}
}
\value{
Returns a modified version of the input data.frame
}
\description{
Takes the mean of the previous and next value for each outlier observation,
and replaces the outlier value with this mean value. Should be used with the
\code{find_outliers()} function.
}
\examples{
silk_data1 |>
  find_outliers("y", group_col = "group") |>
  correct_outliers(col = "y") |>
  head()
}
