\name{make_shift_grid}
\alias{make_shift_grid}
\title{Build Shift Grid}
\usage{make_shift_grid(scenario_name)}
\arguments{\item{scenario_name}{Character string. Scenario name.}}
\value{Numeric vector of shift grid values.}
\description{Creates the sequence of candidate shift values used by registration.}
