% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{method_context}
\alias{method_context}
\title{Create a method context object}
\usage{
method_context(
  method,
  fold_id = 1L,
  replicate_id = 1L,
  n_train_setting = NA,
  time_grid_setting = NA
)
}
\arguments{
\item{method}{Character. Method name.}

\item{fold_id}{Integer. Fold identifier.}

\item{replicate_id}{Integer. Replicate identifier.}

\item{n_train_setting}{Numeric. Training set size setting.}

\item{time_grid_setting}{Character. Time grid setting identifier.}
}
\value{
A list with method context fields.
}
\description{
Create a method context object
}
\keyword{internal}
