% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{predict_residual_cox_risk}
\alias{predict_residual_cox_risk}
\title{Predict risk from a residual Cox model}
\usage{
predict_residual_cox_risk(model, x_new, horizons)
}
\arguments{
\item{model}{Fitted model from \code{fit_residual_cox}.}

\item{x_new}{Numeric matrix of new covariates.}

\item{horizons}{Numeric vector of prediction horizons.}
}
\value{
Matrix of predicted risks (subjects x horizons).
}
\description{
Predict risk from a residual Cox model
}
