\name{pretty_scenario}
\alias{pretty_scenario}
\alias{pretty_method}
\alias{pretty_schedule}
\title{Pretty-Print Labels}
\usage{
pretty_scenario(x)
pretty_method(x)
pretty_schedule(x)
}
\arguments{\item{x}{Character vector of identifiers.}}
\value{Character vector of human-readable labels.}
\description{Convert internal identifiers to human-readable labels for scenarios, methods, and schedules.}
