% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silk-defaults.R
\name{pretty_method}
\alias{pretty_method}
\title{Pretty-print method names}
\usage{
pretty_method(x)
}
\arguments{
\item{x}{Character vector of method identifiers.}
}
\value{
Character vector of human-readable method labels.
}
\description{
Pretty-print method names
}
