% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silk-defaults.R
\name{pretty_scenario}
\alias{pretty_scenario}
\title{Pretty-print scenario names}
\usage{
pretty_scenario(x)
}
\arguments{
\item{x}{Character vector of scenario identifiers.}
}
\value{
Character vector of human-readable scenario labels.
}
\description{
Pretty-print scenario names
}
