% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silk-defaults.R
\name{pretty_schedule}
\alias{pretty_schedule}
\title{Pretty-print schedule names}
\usage{
pretty_schedule(x)
}
\arguments{
\item{x}{Character vector of schedule identifiers.}
}
\value{
Character vector of human-readable schedule labels.
}
\description{
Pretty-print schedule names
}
