% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silk-defaults.R
\name{scenario_schedule}
\alias{scenario_schedule}
\title{Get default schedule for a scenario}
\usage{
scenario_schedule(scenario_name, fallback = NULL)
}
\arguments{
\item{scenario_name}{Character string. A scenario name.}

\item{fallback}{Character string. Fallback schedule name.}
}
\value{
Character string schedule name.
}
\description{
Get default schedule for a scenario
}
\keyword{internal}
