% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silk_data1.R
\docType{data}
\name{silk_data1}
\alias{silk_data1}
\title{Example data for the silk package.}
\format{
\subsection{\code{silk_data1}}{

A data frame with 2,000 rows and 3 columns:
\describe{
\item{time}{x-axis variable, such as time}
\item{y}{Response variable}
\item{group}{Grouping variable for the two series}
...
}
}
}
\source{
silk package
}
\usage{
silk_data1
}
\description{
This dataset is used to demonstrate the functionality of the silk package.
Data contains two series of data, each with 1000 observations. Two outliers
are present in each series.
}
\keyword{datasets}
