% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{step_eval}
\alias{step_eval}
\title{Step-function evaluation}
\usage{
step_eval(times, values, grid)
}
\arguments{
\item{times}{Numeric vector.}

\item{values}{Numeric vector.}

\item{grid}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Step-function evaluation
}
\keyword{internal}
