% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-survival.R
\name{step_eval_cumhaz}
\alias{step_eval_cumhaz}
\title{Evaluate a step-function cumulative hazard on a grid}
\usage{
step_eval_cumhaz(times, cumhaz, grid)
}
\arguments{
\item{times}{Numeric vector of event times.}

\item{cumhaz}{Numeric vector of cumulative hazard values.}

\item{grid}{Numeric vector of evaluation times.}
}
\value{
Numeric vector of cumulative hazard values on the grid.
}
\description{
Evaluate a step-function cumulative hazard on a grid
}
\keyword{internal}
