% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summarize_calibration_bins}
\alias{summarize_calibration_bins}
\title{Summarize calibration bins across simulation replications}
\usage{
summarize_calibration_bins(calibration_bins)
}
\arguments{
\item{calibration_bins}{Data frame returned in the calibration component of
\code{evaluate_prediction_frame}.}
}
\value{
A data frame of weighted calibration-bin summaries.
}
\description{
Summarize calibration bins across simulation replications
}
