% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_silk}
\alias{theme_silk}
\title{SILK ggplot2 theme}
\usage{
theme_silk(base_size = NULL)
}
\arguments{
\item{base_size}{Numeric. Base font size.}
}
\value{
A ggplot2 theme object.
}
\description{
A publication-quality ggplot2 theme for SILK visualizations.
}
