% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-survival.R
\name{trapz}
\alias{trapz}
\title{Trapezoidal integration}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{Numeric vector of grid points.}

\item{y}{Numeric vector of function values.}
}
\value{
Numeric scalar.
}
\description{
Trapezoidal integration
}
\keyword{internal}
