% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers in a dataset}
\usage{
find_outliers(dat, col)
}
\arguments{
\item{dat}{A data.frame with the data}

\item{col}{The column to find outliers from}
}
\value{
A data.frame with something
}
\description{
Find outliers in a dataset
}
\examples{
 # Examples here
}
