% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers in a dataset}
\usage{
find_outliers(x, col, group_col = NULL, threshold = 10)
}
\arguments{
\item{x}{A data.frame with the data}

\item{col}{The column to find outliers from}

\item{group_col}{The column to group by}

\item{threshold}{The threshold value for finding outliers. Outliers are
threshold * MAD away from the median.}
}
\value{
A modified version of the input data.frame, which includes columns for
median, MAD and a logical column to indicate outliers.
}
\description{
A function for finding outliers in a dataset. Uses Median Absolute Deviation
(MAD) to detect values which are a certain distance away from the median
value.
}
\examples{
silk_data1 |>
  find_outliers("y", group_col = "group") |>
  head()
}
